serbot/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.6.0
numpy>=1.23.0
streamlit>=1.30.0
Pillow>=9.0.0
pandas>=1.3.0 